- remove some compiler options and fix resulting build errors


git-svn-id: https://ipvs.informatik.uni-stuttgart.de/SGpp/repos/trunk@3275 4eea3252-f0fb-4393-894d-40516dce545b

+++ src/sgpp/parallel/tools/DynamicTwoPartitionAutoTuning.cpp
@@ -50,7 +50,7 @@ namespace sg {
         // artition 2 that is near 2*_partition2Divider
         // Then, autotuning can continue normally, as there are valid partition2_element_time values.
         if(_partition2_speedup == 0 && _timePartition1/_timePartition2 > 3) {
-          _partition2_speedup = 2.0*_partition2Divider/(_problemSize-_partition2Divider);
+          _partition2_speedup = 2.0*static_cast<double>(_partition2Divider)/static_cast<double>(_problemSize - _partition2Divider);
         }
       }
